+++ apps/server/src/index.ts
@@ -1,27 +1,12 @@
+import corsMiddleware from "@server/middlewares/cors";
 import blogs from "@server/routes/blogs";
 import { Hono } from "hono";
 import { type InferResponseType, hc } from "hono/client";
-import { cors } from "hono/cors";
 
-type Bindings = {
-  CLIENT_URLS: string[];
-  CF_CLIENT_DOMAIN: string;
-};
-
-const app = new Hono<{ Bindings: Bindings }>();
+const app = new Hono();
 
 const router = app
-  .use("*", async (c, next) => {
-    const corsMiddlewareHandler = cors({
-      origin:
-        c.env.CF_CLIENT_DOMAIN !== ""
-          ? (origin) => {
-              return origin.endsWith(c.env.CF_CLIENT_DOMAIN) ? origin : null;
-            }
-          : c.env.CLIENT_URLS,
-    });
-    return corsMiddlewareHandler(c, next);
-  })
+  .use("*", corsMiddleware)
   .get("/", (c) => {
     return c.json({ message: "Hello Hono!" });
   })
